chore: uplift code quality and fix clippy warnings
- Align workspace lints with 2026 Rust best practices
- Fix manual_div_ceil, unlined_format_args, and op_ref warnings
- Ensure all crates inherit workspace lint configuration
- Prepare feat/add-kallsyms-parser for merge to main

+++ crates/kernel_core/src/elf_builder.rs
@@ -95,7 +95,7 @@ impl<'a> ElfBuilder<'a> {
             if sym.sym_type.is_global() { SymbolScope::Dynamic } else { SymbolScope::Compilation };
 
         // Calculate offset within section
-        let offset = sym.address.saturating_sub(self.symbols.kernel_base) as u64;
+        let offset = sym.address.saturating_sub(self.symbols.kernel_base);
 
         obj.add_symbol(Symbol {
             name: sym.name.as_bytes().to_vec(),
